fix pydoc

+++ tensorflow/python/ops/math_ops.py
@@ -4904,22 +4904,21 @@ def sampled_addmm(indices,
   <tf.Tensor: shape=(2,), dtype=int32, numpy=array([2, 2], dtype=int32)>
   >>> mat1 = tf.constant([1, 2, 3, 4, 5, 6], shape=[2, 3], dtype=tf.float32)
   >>> mat1
-  <tf.Tensor: shape=(2, 3), dtype=int32, numpy=
-  array([[1, 2, 3],
-         [4, 5, 6]], dtype=int32)>
+  <tf.Tensor: shape=(2, 3), dtype=float32, numpy=
+  array([[1., 2., 3.],
+         [4., 5., 6.]], dtype=float32)>
   >>> mat2 = tf.constant([7, 8, 9, 10, 11, 12], shape=[3, 2], dtype=tf.float32)
   >>> mat2
-  <tf.Tensor: shape=(3, 2), dtype=int32, numpy=
-  array([[ 7,  8],
-         [ 9, 10],
-         [11, 12]], dtype=int32)>
+  <tf.Tensor: shape=(3, 2), dtype=float32, numpy=
+  array([[ 7.,  8.],
+         [ 9., 10.],
+         [11., 12.]], dtype=float32)>
   >>> tf.sparse.sampled_addmm(indices, values, dense_shape, mat1, mat2,
   ... alpha=0.75, beta=0.25)
   (<tf.Tensor: shape=(2, 2), dtype=int32, numpy=
   array([[0, 0],
-         [1, 1]], dtype=int32)>, 
-  <tf.Tensor: shape=(2,), dtype=float32, numpy=
-  array([ 43.625, 115.575], dtype=float32)>, 
+         [1, 1]], dtype=int32)>, <tf.Tensor: shape=(2,), dtype=float32, numpy=
+  array([ 43.625, 115.575], dtype=float32)>,
   <tf.Tensor: shape=(2,), dtype=int32, numpy=array([2, 2], dtype=int32)>)
 
   A batch operation:
@@ -4929,41 +4928,45 @@ def sampled_addmm(indices,
   <tf.Tensor: shape=(2, 2, 2), dtype=int32, numpy=
   array([[[0, 1],
           [1, 0]],
+
          [[0, 0],
           [1, 0]]], dtype=int32)>
   >>> values = tf.constant([3, 5, 2, 7], shape=[2, 2], dtype=tf.float32)
   >>> values
-  <tf.Tensor: shape=(2, 2), dtype=int32, numpy=
-  array([[3, 5],
-         [2, 7]], dtype=int32)>
+  <tf.Tensor: shape=(2, 2), dtype=float32, numpy=
+  array([[3., 5.],
+         [2., 7.]], dtype=float32)>
   >>> dense_shape = tf.constant([2, 2])
   >>> dense_shape
   <tf.Tensor: shape=(2,), dtype=int32, numpy=array([2, 2], dtype=int32)>
   >>> mat1 = tf.constant(np.arange(1, 13), shape=[2, 2, 3], dtype=tf.float32)
   >>> mat1
-  <tf.Tensor: shape=(2, 2, 3), dtype=int64, numpy=
-  array([[[ 1,  2,  3],
-          [ 4,  5,  6]],
-         [[ 7,  8,  9],
-          [10, 11, 12]]])>
+  <tf.Tensor: shape=(2, 2, 3), dtype=float32, numpy=
+  array([[[ 1.,  2.,  3.],
+          [ 4.,  5.,  6.]],
+
+         [[ 7.,  8.,  9.],
+          [10., 11., 12.]]], dtype=float32)>
   >>> mat2 = tf.constant(np.arange(13, 25), shape=[2, 3, 2], dtype=tf.float32)
   >>> mat2
-  <tf.Tensor: shape=(2, 3, 2), dtype=int64, numpy=
-  array([[[13, 14],
-          [15, 16],
-          [17, 18]],
-         [[19, 20],
-          [21, 22],
-          [23, 24]]])>
+  <tf.Tensor: shape=(2, 3, 2), dtype=float32, numpy=
+  array([[[13., 14.],
+          [15., 16.],
+          [17., 18.]],
+
+         [[19., 20.],
+          [21., 22.],
+          [23., 24.]]], dtype=float32)>
   >>> tf.sparse.sampled_addmm(indices, values, dense_shape, mat1, mat2,
   ... alpha=0.75, beta=0.25)
   (<tf.Tensor: shape=(2, 2, 2), dtype=int32, numpy=
   array([[[0, 1],
           [1, 0]],
+
          [[0, 0],
           [1, 0]]], dtype=int32)>, <tf.Tensor: shape=(2, 2), dtype=float32,
   numpy=array([[ 75.75, 173.  ],
-         [381.5 , 524.5 ]], dtype=float32)>, 
+         [381.5 , 524.5 ]], dtype=float32)>,
   <tf.Tensor: shape=(2,), dtype=int32, numpy=array([2, 2], dtype=int32)>)
 
   Args:
